+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## next
+
+- Added a check for all methods are defined on query compilation to avoid runtime exceptions
+- Improved error message when a methods is not defined, i.e. `Method "foo" is not defined` instead of `m.foo is not a function`
+
 ## 1.0.0-beta.5 (November 10, 2020)
 
 - Jora became a superset of JSON5 (see #23)